requirements.txt
aiohttp==3.8.1
aiohttp-jinja2==1.5
aiosignal==1.2.0
alembic==1.8.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.26.0
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-allauth==0.51.0
djangorestframework==3.13.1
Flask==2.2.2
frozenlist==1.3.1
gino==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.0
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
SQLAlchemy==1.3.24
sqlparse==0.4.2
urllib3==1.26.11
Werkzeug==2.2.2
whitenoise==6.2.0
yarl==1.8.1
zipp==3.8.1
